fix(chat): pin version/credits footer to sidebar bottom
The horizontal sidebar variant was cramming the commit SHAs and
"Made proudly in" credits inline with the waddle switcher buttons, taking
over the top rail on mobile. Extract a VersionFooter component and render
it at the bottom of the mobile nav drawer instead, while the desktop icon
rail keeps the stacked variant at its foot.
