+++ chat/src/components/chat/WaddlesSidebar.vue
@@ -4,8 +4,8 @@ import { Plus, Lock, Compass, MessageCircle } from "lucide-vue-next";
 import type { WaddleSummary } from "@/lib/waddle-api";
 import type { WaddleSession } from "@/lib/server-auth";
 import type { ServerVersion } from "@/composables/useVersion";
-import { extractServerSha } from "@/composables/useVersion";
 import ProfilePanel from "@/components/chat/ProfilePanel.vue";
+import VersionFooter from "@/components/chat/VersionFooter.vue";
 
 const props = defineProps<{
   waddles: WaddleSummary[];
@@ -19,13 +19,6 @@ const props = defineProps<{
   webCommitSha?: string;
   serverVersion?: ServerVersion | null;
 }>();
-
-const webShortSha = computed(() => (props.webCommitSha ?? "unknown").slice(0, 7));
-const serverShortSha = computed(() => {
-  const sha = extractServerSha(props.serverVersion ?? null);
-  if (sha) return sha.slice(0, 7);
-  return props.serverVersion?.version ?? "…";
-});
 
 const emit = defineEmits<{
   selectWaddle: [id: string];
@@ -195,30 +188,11 @@ function waddleColor(waddle: WaddleSummary): string {
       @toggle-notifications="emit('toggle-notifications')"
     />
 
-    <!-- Version / credits footer -->
-    <div
+    <!-- Version / credits footer (vertical only; mobile drawer renders its own) -->
+    <VersionFooter
       v-if="!horizontal"
-      class="flex flex-col items-center gap-0.5 text-[9px] leading-tight text-muted-foreground/60 select-text"
-      :title="`web ${webCommitSha ?? 'unknown'}\nserver ${serverVersion?.version ?? 'unknown'}`"
-    >
-      <span class="font-mono">w {{ webShortSha }}</span>
-      <span class="font-mono">s {{ serverShortSha }}</span>
-      <span class="text-center leading-tight">
-        Made proudly in<br />
-        <span role="img" aria-label="Germany, Norway, and Scotland">🇩🇪🇳🇴🏴󠁧󠁢󠁳󠁣󠁴󠁿</span>
-      </span>
-    </div>
-    <div
-      v-else
-      class="ml-auto flex items-center gap-2 shrink-0 whitespace-nowrap text-[10px] text-muted-foreground/70 select-text"
-      :title="`web ${webCommitSha ?? 'unknown'}\nserver ${serverVersion?.version ?? 'unknown'}`"
-    >
-      <span class="font-mono">web {{ webShortSha }}</span>
-      <span class="font-mono">srv {{ serverShortSha }}</span>
-      <span class="flex items-center gap-1">
-        <span>Made proudly in</span>
-        <span role="img" aria-label="Germany, Norway, and Scotland">🇩🇪🇳🇴🏴󠁧󠁢󠁳󠁣󠁴󠁿</span>
-      </span>
-    </div>
+      :web-commit-sha="webCommitSha"
+      :server-version="serverVersion"
+    />
   </div>
 </template>
